requirements.txt
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0
construct==2.10.68
httpx>=0.23.0,<0.24.0
aiohttp==3.9.1
pydantic==2.5.2
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
tweepy==4.14.0
discord.py==2.3.2
pillow==10.1.0
jinja2==3.1.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
schedule==1.2.0
colorlog==6.8.0
rich==13.7.0
typer==0.9.0
fastapi==0.104.1
uvicorn==0.24.0
